requirements.txt
PyGithub==1.59.1
solana==0.30.2